Skip time when tick is called after closing the menu
Otherwise we attempt to catch up, causing a flood of `update` calls.
This then causes the next `tick` call to be late, especially if the game
does a large amount of work in `update`.

This uses the MENU button release from closing the menu to detect the
menu closing. Could possibly be replaced with a flag somewhere in the
API section (requiring an API minor version update)